feat(canvas): fixes
Ref: ENG-47

Ref: #1040
